+++ libcnmc/cli.py
@@ -1,20 +1,19 @@
 # -*- coding: utf-8 -*-
 import click
 from datetime import datetime
-import os
-import tempfile
 
 from libcnmc.utils import N_PROC
 from libcnmc.core import UpdateCNMCStats, UpdateCINISComptador
 from libcnmc.core.backend import OOOPFactory
 from libcnmc.cir_4_2014 import F1, F1bis, F11
 from libcnmc import cir_3_2015
-from libcnmc import cir_4_2015
+
 from libcnmc.res_4603 import INV
 from libcnmc.res_4603 import CINIMAQ, CINIPOS, CreateCelles, UpdateCINISTrafo
 from libcnmc.res_4603 import UpdateCINISTrams, UpdateCINISCts
 from cli_4666 import cnmc_4666
 from cli_4667 import cnmc_4667
+from cli_4_2015 import cnmc_4_2015
 
 
 @click.group()
@@ -1174,630 +1173,6 @@ def update_cinis_cts(**kwargs):
 @click.option('--interactive/--no-interactive', default=True,
               help="Deshabilitar el mode interactiu")
 @click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f1(**kwargs):
-    """
-    Click entry to generate the F1 of 4/2015
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs: dict(str,str) 
-    :return: None
-    :rtype: None
-    """
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F1(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f11(**kwargs):
-    """
-    Click entry to generate the F11 of 4/2015
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs: dict(str,str)
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F11(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f1bis(**kwargs):
-    """
-    Click entry to gneerate the F1BIS of 4/2015
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs:dict(str, str) 
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F1bis(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f12(**kwargs):
-    """
-    Click entry to generate the F12 of the 4/2015
-    
-    :param kwargs: Params to pass to the proces  
-    :type kwargs: dict(str, str) 
-    :return: None 
-    :rtype: None
-    """
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F12(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f12bis(**kwargs):
-    """
-    Click entry to generate the F12 BIS of the 4/2015
-    
-    :param kwargs: Params to pass to the process 
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None 
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F12bis(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f13(**kwargs):
-    """
-    Click entry to generate the F13 file of 4/2015
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs: dict(str,str) 
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F13(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f13bis(**kwargs):
-    """
-    Click entry to generate the F13 BIS of 4/2015
-    
-    :param kwargs: Params to pas to the process 
-    :type kwargs: dict(str,str) 
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F13bis(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=1,
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f13c(**kwargs):
-    """
-    Click entry to generate the F13C of 4/2015
-    
-    :param kwargs: Params to pass to the process 
-    :type kwargs: dict(str, str) 
-    :return: None
-    :return: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F13c(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f14(**kwargs):
-    """
-    Click entry to generate the F14 file of 4/2015
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs: dict(str,str) 
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F14(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f15(**kwargs):
-    """
-    Click ent ry to generate F15 of 4/2015
-    
-    :param kwargs: Params to pass to the process 
-    :type kwargs: dict(str, str) 
-    :return: None 
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F15(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f10at(**kwargs):
-    """
-    Click entry to gnerate the F10AT of 4/2015
-    
-    :param kwargs: Params to pass to the process 
-    :type kwargs: dict(str, str) 
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F10AT(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f10bt(**kwargs):
-    """
-    Click entry to genertate the F10BT fo 4/2015
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs: dict(str,str)
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F10BT(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f20(**kwargs):
-    """
-    Click entry to generate the F20 file of 4/2015
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs:dict(str,str) 
-    :return: None 
-    :rtype: None
-    """
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F20(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069,
-              help='Port servidor ERP', type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f9(**kwargs):
-    """
-    Click entry to generate the F9 of 4/20105
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs: dict(str, str)
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.F9(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-@click.option('-f',  '--file-input', type=click.Path(exists=True))
-def cir_4_2015_create_celles(**kwargs):
-    """
-    Click entry to create celles of 4/2015
-    
-    :param kwargs: Params to pass to the process
-    :type kwargs: dict(str,str) 
-    :return: None
-    :rtype: None
-    """
-
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-             pwd=kwargs['password'], port=kwargs['port'],
-             uri=kwargs['server'])
-    proc = cir_4_2015.CreateCelles4_2015(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        file_input=kwargs['file_input'])
-    proc.calc()
-
-
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
 @click.option('-y', '--year', default=(datetime.now().year - 1),
               help=u"Any per càlculs")
 @click.option('-s', '--server', default='http://localhost',
@@ -1833,48 +1208,7 @@ def cir_3_2015_f3(**kwargs):
     proc.calc()
 
 
-@cnmc.command()
-@click.option('-q', '--quiet', default=False,
-              help="No mostrar missatges de status per stderr")
-@click.option('--interactive/--no-interactive', default=True,
-              help="Deshabilitar el mode interactiu")
-@click.option('-o', '--output', help="Fitxer de sortida")
-@click.option('-y', '--year', default=(datetime.now().year - 1),
-              help=u"Any per càlculs")
-@click.option('-s', '--server', default='http://localhost',
-              help=u'Adreça servidor ERP')
-@click.option('-p', '--port', default=8069, help='Port servidor ERP',
-              type=click.INT)
-@click.option('-c', '--codi-r1', help='Codi R1 de la distribuidora')
-@click.option('-u', '--user', default='admin', help='Usuari servidor ERP')
-@click.option('-w', '--password', default='admin',
-              help='Contrasenya usuari ERP')
-@click.option('-d', '--database', help='Nom de la base de dades')
-@click.option('--num-proc', default=N_PROC, type=click.INT)
-def cir_4_2015_f16(**kwargs):
-    """
-    Click entry to generate the F16 of 4/2015
-    
-    :param kwargs: Params to pass to the process 
-    :type kwargs: dict(str,str) 
-    :return: None
-    :rtype: None
-    """
-    O = OOOPFactory(dbname=kwargs['database'], user=kwargs['user'],
-                    pwd=kwargs['password'], port=kwargs['port'],
-                    uri=kwargs['server'])
-    proc = cir_4_2015.F16(
-        quiet=kwargs['quiet'],
-        interactive=kwargs['interactive'],
-        output=kwargs['output'],
-        connection=O,
-        num_proc=kwargs['num_proc'],
-        codi_r1=kwargs['codi_r1'],
-        year=kwargs['year']
-    )
-    proc.calc()
-
-cli = click.CommandCollection(sources=[cnmc, cnmc_4666, cnmc_4667])
+cli = click.CommandCollection(sources=[cnmc, cnmc_4666, cnmc_4667, cnmc_4_2015])
 
 if __name__ == '__main__':
     cli()
